refactor: orderPriceSummary 유틸 함수를 컴포넌트로 이동
