+++ src/components/features/coupon/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useAppliedCoupons';
-export * from './useOrderSummary';
